embed get-plugin.js
